pyproject.toml
[tool.poetry]
name = "boilerplate_fast_api"
version = "0.1.0"
description = ""
authors = ["DiegoBulhoes <diegobulhoesm@gmail.com>", "ViniciusEspindola <vinnyaoe@gmail.com>"]
readme = "README.md"
packages = [{include = "boilerplate_fastapi"}]

[tool.poetry.dependencies]
python = "^3.8.10"
uvicorn = "^0.18.3"
fastapi = "^0.83.0"
python-decouple = "^3.8"
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
pyjwt = "^2.6.0"
pytest-mock = "^3.10.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
black = "^22.8.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
